requirements.txt
numpy>=1.18.3
pandas>=1.0.3
confuse>=1.3.0
pyarrow>=0.16.0
tabulate>=0.8.7
codetiming>=1.2.0
holidays>=0.10.2
requests>=2.23.0
Flask~=1.1.2
Flask-HTTPAuth~=4.0.0
Flask-Mail~=0.9.1
SQLAlchemy~=1.3.18
Flask-Caching~=1.9.0
google-cloud-firestore>=1.7.0
google-cloud-bigquery>=1.25.0
google-cloud-storage>=1.29.0
apache-beam==2.21.0
apache-beam[gcp]
